form array aot build fix (#884)
