+++ package.json
@@ -39,50 +39,50 @@
   },
   "homepage": "https://github.com/udos86/ng-dynamic-forms#readme",
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/core": "^7.1.4",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/material": "^7.2.0",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
-    "@ionic/angular": "^4.0.0-rc.0",
+    "@angular/animations": "^7.2.1",
+    "@angular/cdk": "^7.2.1",
+    "@angular/common": "^7.2.1",
+    "@angular/compiler": "^7.2.1",
+    "@angular/compiler-cli": "^7.2.1",
+    "@angular/core": "^7.2.1",
+    "@angular/forms": "^7.2.1",
+    "@angular/http": "^7.2.1",
+    "@angular/material": "^7.2.1",
+    "@angular/platform-browser": "^7.2.1",
+    "@angular/platform-browser-dynamic": "^7.2.1",
+    "@angular/router": "^7.2.1",
+    "@ionic/angular": "^4.0.0-rc.2",
     "@ng-bootstrap/ng-bootstrap": "^4.0.1",
-    "@progress/kendo-angular-dateinputs": "^3.5.2",
-    "@progress/kendo-angular-dropdowns": "^3.4.1",
-    "@progress/kendo-angular-inputs": "^3.5.0",
+    "@progress/kendo-angular-dateinputs": "^3.5.4",
+    "@progress/kendo-angular-dropdowns": "^3.4.2",
+    "@progress/kendo-angular-inputs": "^4.0.0",
     "@progress/kendo-angular-intl": "^1.6.1",
     "@progress/kendo-angular-l10n": "^1.3.0",
     "@progress/kendo-angular-popup": "^2.5.0",
     "@progress/kendo-angular-resize-sensor": "^3.1.2",
     "@progress/kendo-angular-upload": "^4.2.2",
     "@progress/kendo-common": "^0.1.1",
-    "@progress/kendo-date-math": "^1.3.2",
-    "@progress/kendo-popup-common": "^1.6.0",
-    "@progress/kendo-theme-default": "^2.62.0",
+    "@progress/kendo-date-math": "^1.4.0",
+    "@progress/kendo-popup-common": "^1.6.1",
+    "@progress/kendo-theme-default": "^3.2.0",
     "@telerik/kendo-intl": "^1.4.4",
     "angular2-text-mask": "^9.0.0",
     "bootstrap": "^3.4.0",
-    "core-js": "^2.6.1",
+    "core-js": "^2.6.2",
     "foundation-sites": "^6.5.1",
     "hammerjs": "^2.0.8",
     "ngx-bootstrap": "^3.1.4",
     "primeicons": "^1.0.0",
-    "primeng": "^7.0.3",
+    "primeng": "^7.0.4",
     "quill": "^1.3.6",
     "rxjs": "^6.3.3",
     "rxjs-compat": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.8.28"
   },
   "devDependencies": {
-    "@ngtools/webpack": "^7.1.4",
+    "@ngtools/webpack": "^7.2.2",
     "@types/hammerjs": "^2.0.36",
-    "@types/jasmine": "^3.3.5",
+    "@types/jasmine": "^3.3.7",
     "@types/node": "^10.12.18",
     "angular-router-loader": "^0.8.5",
     "angular2-template-loader": "^0.6.2",
@@ -114,26 +114,26 @@
     "merge-stream": "^1.0.1",
     "protractor": "^5.4.2",
     "raw-loader": "^1.0.0",
-    "rollup": "^1.0.0",
+    "rollup": "^1.1.0",
     "rollup-plugin-alias": "^1.5.1",
     "rollup-plugin-commonjs": "^9.2.0",
     "rollup-plugin-node-resolve": "^4.0.0",
     "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-uglify": "^6.0.0",
+    "rollup-plugin-uglify": "^6.0.1",
     "rollup-stream": "^1.24.1",
     "run-sequence": "^2.2.1",
     "systemjs": "0.21.5",
-    "ts-loader": "^5.3.2",
+    "ts-loader": "^5.3.3",
     "tsickle": "^0.33.1",
-    "tslint": "^5.12.0",
+    "tslint": "^5.12.1",
     "typedoc": "^0.13.0",
     "typescript": "^3.1.6",
     "uglify-es": "^3.3.9",
     "vinyl-buffer": "^1.0.1",
     "vinyl-source-stream": "^2.0.0",
-    "webpack": "^4.28.3",
-    "webpack-cli": "^3.1.2",
+    "webpack": "^4.28.4",
+    "webpack-cli": "^3.2.1",
     "webpack-dev-server": "^3.1.14",
-    "webpack-merge": "^4.1.5"
+    "webpack-merge": "^4.2.1"
   }
 }
