+++ packages/core/src/component/dynamic-form-array.component.ts
@@ -1,4 +1,5 @@
 import { QueryList } from "@angular/core";
+import { FormArray } from "@angular/forms";
 import { DynamicFormControlComponent } from "./dynamic-form-control.component";
 import { DynamicTemplateDirective } from "../directive/dynamic-template.directive";
 import { DynamicFormArrayModel } from "../model/form-array/dynamic-form-array.model";
@@ -7,6 +8,10 @@ export abstract class DynamicFormArrayComponent extends DynamicFormControlCompon
 
     model: DynamicFormArrayModel;
     templates: QueryList<DynamicTemplateDirective> | undefined;
+
+    get array(): FormArray {
+        return this.control as FormArray;
+    }
 
     get startTemplate(): DynamicTemplateDirective | undefined {
         return this.layoutService.getStartTemplate(this.model, this.templates);
