+++ packages/core/src/core.ts
@@ -49,6 +49,8 @@ export * from "./service/dynamic-form-layout.service";
 export * from "./service/dynamic-form-validation.service";
 export * from "./service/dynamic-form-instances.service";
 
+export * from "./service/dynamic-form.validators";
+
 export * from "./utils/autofill.utils";
 export * from "./utils/core.utils";
 export * from "./utils/json.utils";
