+++ packages/ui-bootstrap/src/form-array/dynamic-bootstrap-form-array.component.html
@@ -9,7 +9,7 @@
 
             <dynamic-bootstrap-form-control *ngFor="let _model of groupModel.group"
                                             [context]="groupModel"
-                                            [group]="control.at(idx)"
+                                            [group]="array.at(idx)"
                                             [hidden]="_model.hidden"
                                             [layout]="layout"
                                             [model]="_model"
