+++ packages/ui-ng-bootstrap/src/form-array/dynamic-ng-bootstrap-form-array.component.html
@@ -9,7 +9,7 @@
 
             <dynamic-ng-bootstrap-form-control *ngFor="let _model of groupModel.group"
                                                [context]="groupModel"
-                                               [group]="control.at(idx)"
+                                               [group]="array.at(idx)"
                                                [hidden]="_model.hidden"
                                                [layout]="layout"
                                                [model]="_model"
