+++ packages/ui-primeng/src/form-array/dynamic-primeng-form-array.component.html
@@ -11,7 +11,7 @@
 
                 <dynamic-primeng-form-control *ngFor="let _model of groupModel.group"
                                               [context]="groupModel"
-                                              [group]="control.at(idx)"
+                                              [group]="array.at(idx)"
                                               [hidden]="_model.hidden"
                                               [layout]="layout"
                                               [model]="_model"
